refactor: 알림카드에 체크 버튼 주입하는 컴포넌트 합성 방식으로 구조 변경 [BETA-113]
